# Add adler32 SIMD ARM support