--- a/adler32.c
+++ b/adler32.c
@@ -61,6 +61,10 @@ local uLong adler32_combine_ OF((uLong adler1, uLong adler2, z_off64_t len2));
 #  define MOD63(a) a %= BASE
 #endif
 
+#if defined(ADLER32_SIMD_NEON)
+#include "adler32_simd.h"
+#endif
+
 /* ========================================================================= */
 uLong ZEXPORT adler32(adler, buf, len)
     uLong adler;
@@ -69,6 +73,11 @@ uLong ZEXPORT adler32(adler, buf, len)
 {
     unsigned long sum2;
     unsigned n;
+
+#if defined(ADLER32_SIMD_NEON)
+    if (buf && len >= 64)
+        return adler32_simd_(adler, buf, len);
+#endif
 
     /* split Adler-32 into component sums */
     sum2 = (adler >> 16) & 0xffff;