@janaknat 2020-08-10
Some inflate performance number breakdown (average of 5 runs):
Dataset| File | Throughput before (kb/ms)| Throughput after (kb/ms)| % change |
--------|-----|--------------------|----------------|---|
Silesia | xml | 63.50 | 72.09| +13%
 | | x-ray | 129.93 | 135.26| +4%
 | | webster | 77.89 | 83.48| +7%
 | | sao | 144.97 | 156.53 | +8%
 | | samba | 85.39 | 93.70 | +10%
 | | reymont | 77.01 | 83.29 | +8%
 | | osdb | 99.57 | 110.51 | +11%
 | | ooffice | 100.62 | 108.32 | +8%
 | | nci | 53.23 | 61.98 | +16%
 | | mr | 90.40 | 97.72 | +8%
 | | mozilla | 99.64 | 107.76 | +8%
 | | dickens | 85.06 | 91.04 | +7%